+++ src/Berthe/DAL/Writer.php
@@ -17,14 +17,14 @@ interface Writer {
      * @return boolean
      */
     public function update(\Berthe\VO $object);
-    
+
     /**
      * Delete the object from database
      * @param \Berthe\VO $object the object to insert
      * @return boolean
      */
     public function delete(\Berthe\VO $object);
-    
+
     /**
      * Delete an object by id from database
      * @param int $int object identifier
